feat: add ability to customize metrics port
This change allows a user to specify a metrics port as part of their `FluentBit` custom resource.  This is useful for those who need to use a port other than 2020.

An example of a partial FluentBit resource with this new property might look like this:

```
apiVersion: fluentbit.fluent.io/v1alpha2
kind: FluentBit
metadata:
  name: fluent-bit
  namespace: fluent
  labels:
    app.kubernetes.io/name: fluent-bit
spec:
  image: kubesphere/fluent-bit:v2.0.9
  metricsPort: 32020
.
.
.
```

Signed-off-by: Dale Hille <dh185221@ncr.com>